+++ plugins/main/public/components/common/tables/table-wz-api.tsx
@@ -26,8 +26,7 @@ import { TableWithSearchBar } from './table-with-search-bar';
 import { TableDefault } from './table-default';
 import { WzRequest } from '../../../react-services/wz-request';
 import { ExportTableCsv } from './components/export-table-csv';
-import { useStateStorage } from '../hooks';
-
+import { useStateStorage, useAppConfig } from '../hooks';
 /**
  * Search input custom filter button
  */
@@ -37,9 +36,20 @@ interface CustomFilterButton {
   value: string;
 }
 
-const getFilters = filters => {
+interface Filters {
+  [key: string]: string;
+}
+
+const getFilters = (filters: Filters) => {
   const { default: defaultFilters, ...restFilters } = filters;
   return Object.keys(restFilters).length ? restFilters : defaultFilters;
+};
+
+const formatSorting = sorting => {
+  if (!sorting.field || !sorting.direction) {
+    return '';
+  }
+  return `${sorting.direction === 'asc' ? '+' : '-'}${sorting.field}`;
 };
 
 export function TableWzAPI({
@@ -53,12 +63,11 @@ export function TableWzAPI({
   actionButtons?:
     | ReactNode
     | ReactNode[]
-    | (({ filters }: { filters }) => ReactNode);
+    | (({ filters }: { filters: Filters }) => ReactNode);
   postActionButtons?:
     | ReactNode
     | ReactNode[]
-    | (({ filters }: { filters }) => ReactNode);
-
+    | (({ filters }: { filters: Filters }) => ReactNode);
   title?: string;
   addOnTitle?: ReactNode;
   description?: string;
@@ -67,7 +76,7 @@ export function TableWzAPI({
   searchTable?: boolean;
   endpoint: string;
   buttonOptions?: CustomFilterButton[];
-  onFiltersChange?: Function;
+  onFiltersChange?: (filters: Filters) => void;
   showReload?: boolean;
   searchBarProps?: any;
   reload?: boolean;
@@ -75,10 +84,10 @@ export function TableWzAPI({
   setReload?: (newValue: number) => void;
 }) {
   const [totalItems, setTotalItems] = useState(0);
-  const [filters, setFilters] = useState({});
+  const [filters, setFilters] = useState<Filters>({});
   const [isLoading, setIsLoading] = useState(false);
-
-  const onFiltersChange = filters =>
+  const [sort, setSort] = useState({});
+  const onFiltersChange = (filters: Filters) =>
     typeof rest.onFiltersChange === 'function'
       ? rest.onFiltersChange(filters)
       : null;
@@ -101,16 +110,17 @@ export function TableWzAPI({
       : undefined,
   );
   const [isOpenFieldSelector, setIsOpenFieldSelector] = useState(false);
-
+  const appConfig = useAppConfig();
+  const maxRows = appConfig.data['reports.csv.maxRows'];
   const onSearch = useCallback(async function (
     endpoint,
-    filters,
+    filters: Filters,
     pagination,
     sorting,
   ) {
     try {
       const { pageIndex, pageSize } = pagination;
-      const { field, direction } = sorting.sort;
+      setSort(sorting.sort);
       setIsLoading(true);
       setFilters(filters);
       onFiltersChange(filters);
@@ -118,9 +128,8 @@ export function TableWzAPI({
         ...getFilters(filters),
         offset: pageIndex * pageSize,
         limit: pageSize,
-        sort: `${direction === 'asc' ? '+' : '-'}${field}`,
+        sort: formatSorting(sorting.sort),
       };
-
       const response = await WzRequest.apiReq('GET', endpoint, { params });
 
       const { affected_items: items, total_affected_items: totalItems } = (
@@ -182,7 +191,9 @@ export function TableWzAPI({
   };
 
   useEffect(() => {
-    if (rest.reload) triggerReload();
+    if (rest.reload) {
+      triggerReload();
+    }
   }, [rest.reload]);
 
   const ReloadButton = (
@@ -227,16 +238,22 @@ export function TableWzAPI({
             {rest.showReload && ReloadButton}
             {/* Render optional export to CSV button */}
             {rest.downloadCsv && (
-              <ExportTableCsv
-                endpoint={rest.endpoint}
-                totalItems={totalItems}
-                filters={getFilters(filters)}
-                title={
-                  typeof rest.downloadCsv === 'string'
-                    ? rest.downloadCsv
-                    : rest.title
-                }
-              />
+              <>
+                <ExportTableCsv
+                  endpoint={rest.endpoint}
+                  totalItems={totalItems}
+                  filters={getFilters({
+                    ...filters,
+                    sort: formatSorting(sort),
+                  })}
+                  title={
+                    typeof rest.downloadCsv === 'string'
+                      ? rest.downloadCsv
+                      : rest.title
+                  }
+                  maxRows={maxRows}
+                />
+              </>
             )}
             {/* Render optional post custom action button */}
             {renderActionButtons(postActionButtons, filters)}